+++ tracer/src/Datadog.Tracer.Native/debugger_members.h
@@ -165,7 +165,8 @@ enum class ProbeStatus
      * \brief Preceding with underscore because ERROR is a widely used preprocessor constant.
      */
     // ReSharper disable once CppInconsistentNaming
-    _ERROR  // NOLINT(clang-diagnostic-reserved-identifier, bugprone-reserved-identifier)
+    _ERROR,  // NOLINT(clang-diagnostic-reserved-identifier, bugprone-reserved-identifier)
+    INSTRUMENTED
 };
 
 struct ProbeMetadata
